feat(meeting): implement soft delete for meeting records
add soft delete functionality to meeting records and restructure DTOs into separate files
update API to return deleted record information
add deletedAt field filtering to repository queries
