+++ src/meeting/dto/index.ts
@@ -1,0 +1,6 @@
+export * from './meeting-record-query.dto';
+export * from './meeting-record-response.dto';
+export * from './meeting-record-stats.dto';
+export * from './meeting-record-delete.dto';
+export * from './meeting-record-create.dto';
+export * from './meeting-record-update.dto';
